+++ sumologic/provider.go
@@ -107,6 +107,7 @@ func Provider() terraform.ResourceProvider {
 			"sumologic_policies":                                 resourceSumologicPolicies(),
 			"sumologic_hierarchy":                                resourceSumologicHierarchy(),
 			"sumologic_content_permission":                       resourceSumologicPermissions(),
+			"sumologic_local_file_source":                        resourceSumologicLocalFileSource(),
 			"sumologic_log_search":                               resourceSumologicLogSearch(),
 		},
 		DataSourcesMap: map[string]*schema.Resource{