+++ sumologic/resource_sumologic_monitors_library_monitor_test.go
@@ -2,6 +2,8 @@ package sumologic
 
 import (
 	"fmt"
+	"github.com/google/go-cmp/cmp"
+	"github.com/google/go-cmp/cmp/cmpopts"
 	"log"
 	"regexp"
 	"sort"
@@ -65,6 +67,22 @@ func TestSumologicMonitorsLibraryMonitor_conversionsToFromTriggerConditionsShoul
 			exampleMetricsMissingDataTriggerCondition("MissingData"),
 			exampleMetricsMissingDataTriggerCondition("ResolvedMissingData"),
 		},
+		{
+			exampleSloSliTriggerCondition("Critical", 90),
+			exampleSloSliTriggerCondition("Warning", 95),
+		},
+		{
+			exampleSloBurnRateTriggerCondition("Critical"),
+			exampleSloBurnRateTriggerCondition("Warning"),
+		},
+		{
+			exampleSloBurnRateTriggerConditionWithOnlyBurnRates("Critical"),
+			exampleSloBurnRateTriggerConditionWithOnlyBurnRates("Warning"),
+		},
+		{
+			exampleSloBurnRateTriggerConditionWithoutBurnRates("Critical"),
+			exampleSloBurnRateTriggerConditionWithoutBurnRates("Warning"),
+		},
 	}
 	for _, triggerConditions := range testTriggerConditions {
 		triggerConditionsAfterRoundTrip := triggerConditionsBlockToJson(jsonToTriggerConditionsBlock(triggerConditions))
@@ -74,7 +92,14 @@ func TestSumologicMonitorsLibraryMonitor_conversionsToFromTriggerConditionsShoul
 			log.Fatalln("Test case:", triggerConditions, "Lengths differ: Expected", len(triggerConditions), "got", len(triggerConditionsAfterRoundTrip))
 		}
 		for i := range triggerConditions {
-			if triggerConditionsAfterRoundTrip[i] != triggerConditions[i] {
+			if !cmp.Equal(triggerConditionsAfterRoundTrip[i], triggerConditions[i],
+				cmpopts.AcyclicTransformer("s", func(t TriggerCondition) TriggerCondition { //replacing empty burnRates with nil for equality purpose
+					if len(t.BurnRates) == 0 {
+						t.BurnRates = nil
+						return t
+					}
+					return t
+				})) {
 				log.Fatalln("Test case:", triggerConditions, "Expected", triggerConditions[i], "got", triggerConditionsAfterRoundTrip[i])
 			}
 		}
@@ -612,6 +637,42 @@ func TestAccSumologicMonitorsLibraryMonitor_folder_update(t *testing.T) {
 	})
 }
 
+func TestAccSumologicMonitorsLibraryMonitor_override_payload(t *testing.T) {
+	var monitorsLibraryMonitor MonitorsLibraryMonitor
+	testNameSuffix := acctest.RandString(16)
+
+	testName := "terraform_test_monitor_connection_" + testNameSuffix
+	testType := "MonitorsLibraryMonitor"
+	testMonitorType := "Logs"
+	defaultPayload := "{\"eventType\" : \"{{Name}}\"}"
+	resolutionPayload := "{\"eventType\" : \"{{Name}}\"}"
+
+	overrideDefaultPayload := "{\"eventType\" : \"{{Name}}-update\"}"
+	overrideResolutionPayload := "{\"eventType\" : \"{{Name}}-update\"}"
+
+	resource.Test(t, resource.TestCase{
+		PreCheck:     func() { testAccPreCheck(t) },
+		Providers:    testAccProviders,
+		CheckDestroy: testAccCheckMonitorsLibraryMonitorDestroy(monitorsLibraryMonitor),
+		//destroy conection too
+		Steps: []resource.TestStep{
+			{
+				Config: testAccSumologicMonitorsLibraryMonitorUpdateConection(testNameSuffix, defaultPayload, resolutionPayload,
+					overrideDefaultPayload, overrideResolutionPayload),
+				Check: resource.ComposeTestCheckFunc(
+					testAccCheckMonitorsLibraryMonitorExists("sumologic_monitor.test_monitor_connection", &monitorsLibraryMonitor, t),
+					testAccCheckMonitorsLibraryMonitorAttributes("sumologic_monitor.test_monitor_connection"),
+					resource.TestCheckResourceAttr("sumologic_monitor.test_monitor_connection", "monitor_type", testMonitorType),
+					resource.TestCheckResourceAttr("sumologic_monitor.test_monitor_connection", "name", testName),
+					resource.TestCheckResourceAttr("sumologic_monitor.test_monitor_connection", "type", testType),
+					resource.TestCheckResourceAttr("sumologic_monitor.test_monitor_connection", "notifications.0.notification.0.payload_override", overrideDefaultPayload+"\n"),
+					resource.TestCheckResourceAttr("sumologic_monitor.test_monitor_connection", "notifications.0.notification.0.resolution_payload_override", overrideResolutionPayload+"\n"),
+				),
+			},
+		},
+	})
+}
+
 func testAccCheckMonitorsLibraryMonitorFolderMatch(monitorName string, folderName string, t *testing.T) resource.TestCheckFunc {
 	return func(s *terraform.State) error {
 		//fetching monitor information
@@ -913,6 +974,87 @@ resource "sumologic_role" "tf_test_role_01" {
 }`, testName, testName)
 }
 
+func testAccSumologicMonitorsLibraryMonitorUpdateConection(testName string, defaultPayload string, resolutionPayload string,
+	overrideDefaultPayload string, overrideResolutionPayload string) string {
+	return fmt.Sprintf(`
+resource "sumologic_connection" "connection_01" {
+		name = "%s"
+		type = "WebhookConnection"
+		description = "WebhookConnection"
+		url = "https://example.com"
+		webhook_type = "Webhook"
+		default_payload = <<JSON
+%s
+	JSON
+		resolution_payload = <<JSON
+%s
+	JSON
+}
+
+resource "sumologic_monitor" "test_monitor_connection" {
+	name = "terraform_test_monitor_connection_%s"
+	description = "terraform_test_monitor_connection"
+	type = "MonitorsLibraryMonitor"
+	is_disabled = true
+	content_type = "Monitor"
+	monitor_type = "Logs"
+	evaluation_delay = "8m"
+	queries {
+		row_id = "A"
+		query = "_sourceCategory=monitor-manager info"
+	  }
+	triggers  {
+		threshold_type = "GreaterThan"
+		threshold = 40.0
+		time_range = "30m"
+		occurrence_type = "ResultCount"
+		trigger_source = "AllResults"
+		trigger_type = "Critical"
+		detection_method = "StaticCondition"
+	  }
+	  triggers  {
+		threshold_type = "LessThanOrEqual"
+		threshold = 40.0
+		time_range = "30m"
+		occurrence_type = "ResultCount"
+		trigger_source = "AllResults"
+		trigger_type = "ResolvedCritical"
+		detection_method = "StaticCondition"
+		resolution_window = "15m"
+	  }
+	notifications {
+		notification {
+			connection_type = "Webhook"			
+			connection_id = sumologic_connection.connection_01.id
+			payload_override = <<JSON
+%s
+			JSON
+			resolution_payload_override = <<JSON
+%s
+			JSON
+		}
+		run_for_trigger_types = ["Critical", "ResolvedCritical"]
+	}
+	playbook = "This is an updated test playbook"
+	alert_name = "Updated Alert from {{Name}}"
+	
+	obj_permission {
+		subject_type = "role"
+		subject_id = sumologic_role.tf_test_role_01.id
+		permissions = ["Read","Update"]
+	}
+}
+resource "sumologic_role" "tf_test_role_01" {
+	name        = "tf_test_role_01_%s"
+	description = "Testing resource sumologic_role"
+	capabilities = [
+		"viewAlerts",
+		"viewMonitorsV2",
+		"manageMonitorsV2"
+	]
+}`, testName, defaultPayload, resolutionPayload, testName, overrideDefaultPayload, overrideResolutionPayload, testName)
+}
+
 func testAccEmulateFGPDriftingMonitor(
 	t *testing.T,
 	// expectedFGPFunc func(*terraform.State, string) ([]CmfFgpPermStatement, error),
@@ -1077,6 +1219,70 @@ resource "sumologic_monitor" "test" {
 	return resourceText
 }
 
+func exampleSloMonitorWithTriggerCondition(
+	testName string,
+	trigger string) string {
+	var resourceText = fmt.Sprintf(`
+resource "sumologic_slo_folder" "tf_slo_folder" {
+  name        = "slo-tf-folder"
+  description = "folder for SLO created for testing"
+}
+
+resource "sumologic_slo" "slo_tf_window_metric_ratio" {
+  name        = "%s"
+  description = "per minute login error rate over rolling 7 days"
+  parent_id   = sumologic_slo_folder.tf_slo_folder.id
+  signal_type = "Error"
+  service     = "auth"
+  application = "login"
+  compliance {
+      compliance_type = "Rolling"
+      size            = "7d"
+      target          = 95
+      timezone        = "Asia/Kolkata"
+  }
+  indicator {
+    window_based_evaluation {
+      op         = "LessThan"
+      query_type = "Metrics"
+      size       = "1m"
+      threshold  = 99.0
+      queries {
+        query_group_type = "Unsuccessful"
+        query_group {
+          row_id        = "A"
+          query         = "service=auth api=login metric=HTTP_5XX_Count"
+          use_row_count = false
+        }
+      }
+      queries {
+        query_group_type = "Total"
+        query_group {
+          row_id = "A"
+          query  = "service=auth api=login metric=TotalRequests"
+          use_row_count = false
+        }
+      }
+    }
+  }
+}
+
+resource "sumologic_monitor" "test" {
+	name = "%s"
+	description = "terraform_test_monitor_description"
+	type = "MonitorsLibraryMonitor"
+	is_disabled = false
+	content_type = "Monitor"
+	monitor_type = "Slo"
+	slo_id = sumologic_slo.slo_tf_window_metric_ratio.id
+    trigger_conditions {
+      %s
+    }
+	playbook = "This is a test playbook"
+}`, testName, testName, trigger)
+	return resourceText
+}
+
 var exampleLogsStaticTriggerConditionBlock = `
    logs_static_condition {
      critical {
@@ -1177,6 +1383,27 @@ var exampleMetricsMissingDataTriggerConditionBlock = `
      trigger_source = "AnyTimeSeries"
    }`
 
+var exampleSloSliTriggerConditionBlock = `
+   slo_sli_condition {
+	critical {
+		sli_threshold = 90
+	}
+   }`
+
+var exampleSloBurnRateTriggerConditionBlock = `
+   slo_burn_rate_condition {
+      critical {
+        burn_rate_threshold =  20
+        time_range = "1d"
+      }
+      warning {
+        burn_rate {
+             burn_rate_threshold =  20
+             time_range = "2d"
+        }
+      }
+    }`
+
 func exampleLogsStaticMonitor(testName string) string {
 	query := "error | timeslice 1m | count as field by _timeslice"
 	return exampleMonitorWithTriggerCondition(testName, "Logs", query,
@@ -1225,6 +1452,14 @@ func exampleMetricsMissingDataMonitor(testName string) string {
 		exampleMetricsMissingDataTriggerConditionBlock, []string{"MissingData", "ResolvedMissingData"})
 }
 
+func exampleSloSliMonitor(testName string) string {
+	return exampleSloMonitorWithTriggerCondition(testName, exampleSloSliTriggerConditionBlock)
+}
+
+func exampleSloBurnRateMonitor(testName string) string {
+	return exampleSloMonitorWithTriggerCondition(testName, exampleSloBurnRateTriggerConditionBlock)
+}
+
 var allExampleMonitors = []func(testName string) string{
 	exampleLogsStaticMonitor,
 	exampleLogsStaticMonitorWithResolutionWindow,
@@ -1234,6 +1469,8 @@ var allExampleMonitors = []func(testName string) string{
 	exampleMetricsOutlierMonitor,
 	exampleLogsMissingDataMonitor,
 	exampleMetricsMissingDataMonitor,
+	exampleSloSliMonitor,
+	exampleSloBurnRateMonitor,
 }
 
 func testAccSumologicMonitorsLibraryMonitorWithInvalidTriggerCondition(testName string, triggerCondition string) string {
@@ -1315,6 +1552,41 @@ func exampleMetricsMissingDataTriggerCondition(triggerType string) TriggerCondit
 		TriggerType:     triggerType,
 		TriggerSource:   "AllTimeSeries",
 		DetectionMethod: "MetricsMissingDataCondition",
+	}
+}
+
+func exampleSloSliTriggerCondition(triggerType string, threshold float64) TriggerCondition {
+	return TriggerCondition{
+		SLIThreshold:    threshold,
+		TriggerType:     triggerType,
+		DetectionMethod: "SloSliCondition",
+	}
+}
+
+func exampleSloBurnRateTriggerConditionWithOnlyBurnRates(triggerType string) TriggerCondition {
+	return TriggerCondition{
+		TriggerType:     triggerType,
+		DetectionMethod: "SloBurnRateCondition",
+		BurnRates:       []BurnRate{{BurnRateThreshold: 10, TimeRange: "1h"}, {BurnRateThreshold: 30, TimeRange: "1d"}},
+	}
+}
+
+func exampleSloBurnRateTriggerConditionWithoutBurnRates(triggerType string) TriggerCondition {
+	return TriggerCondition{
+		TimeRange:         "30m",
+		BurnRateThreshold: 10,
+		TriggerType:       triggerType,
+		DetectionMethod:   "SloBurnRateCondition",
+	}
+}
+
+func exampleSloBurnRateTriggerCondition(triggerType string) TriggerCondition {
+	return TriggerCondition{
+		TimeRange:         "30m",
+		BurnRateThreshold: 60,
+		TriggerType:       triggerType,
+		DetectionMethod:   "SloBurnRateCondition",
+		BurnRates:         []BurnRate{{BurnRateThreshold: 10, TimeRange: "1h"}, {BurnRateThreshold: 30, TimeRange: "1d"}},
 	}
 }
 
@@ -1400,6 +1672,8 @@ var allInvalidTriggerConditionMonitorResources = []func(testName string) string{
 	invalidExampleWithNoTriggerCondition,
 	invalidExampleWithEmptyLogStaticTriggerCondition,
 	invalidExampleWithEmptyMetricsStaticTriggerCondition,
+	invalidExampleWithEmptySloSliTriggerCondition,
+	invalidExampleWithEmptySloSliTriggerCondition,
 }
 
 func invalidExampleWithNoTriggerCondition(testName string) string {
@@ -1416,4 +1690,12 @@ func invalidExampleWithEmptyMetricsStaticTriggerCondition(testName string) strin
 	query := "error | timeslice 1m | count as field by _timeslice"
 	return exampleMonitorWithTriggerCondition(testName, "Logs", query,
 		`metrics_static_condition {}`, []string{"Critical", "ResolvedCritical"})
+}
+
+func invalidExampleWithEmptySloSliTriggerCondition(testName string) string {
+	return exampleSloMonitorWithTriggerCondition(testName, `slo_sli_condition {}`)
+}
+
+func invalidExampleWithEmptySloBurnRateTriggerCondition(testName string) string {
+	return exampleSloMonitorWithTriggerCondition(testName, `slo_burn_rate_condition {}`)
 }